UPDATE SIMPLE CRUD
SIMPLE CRUD semoga tidak ada yang ngebug sama sekali
